% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocuskeletoncreator.R
\name{add_dcf_info}
\alias{add_dcf_info}
\title{Adding DESCRIPTION information to an existing package}
\usage{
add_dcf_info(psPkgPath, psDescTitle = NULL, psDescAuthor = NULL,
  psDescMaintainer = NULL, psDescDescription = NULL, psDescLicense = NULL)
}
\arguments{
\item{psPkgPath}{package path}

\item{psDescTitle}{title of the package}

\item{psDescAuthor}{author information}

\item{psDescDescription}{Description field in dcf}

\item{psDescLicense}{license information}
}
\description{
This function \code{add_dcf_info} uses the reference class
\code{DcfRefClass} to represent the information stored in
a DESCRIPTION file. Any arguments specified to \code{add_dcf_info}
that are not null, are written to the DESCRIPTION file
and therebye overwritting the existing information in the
DESCRIPTION file.
}

