% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocucleanuptools.R
\name{cleanup_vignettes}
\alias{cleanup_vignettes}
\title{Cleaning up a vignette directory}
\usage{
cleanup_vignettes(psPath = "rmd", psPattern = "pdf$")
}
\arguments{
\item{psPattern}{that files to be clean match}

\item{pkg}{package directory}
}
\description{
\code{cleanup_vignettes} removes all output files that were
produced by any tools. The goal is to have only source files
after the cleanup process.
}

