% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocugraphicsconverter.R
\name{convertLibOToPdf}
\alias{convertLibOToPdf}
\title{Convert documents from source format into target format used in the document}
\usage{
convertLibOToPdf(psLibOFile, psLibODir = "odg", psFigOutDir = ".")
}
\arguments{
\item{psLibOFile}{name of the libre office graphics file}

\item{psLibODir}{source directory of Libre Office files}

\item{psFigOutDir}{output directory where figure pdfs are expected to be}
}
\description{
\code{convertLibOToPdf} assumes that graphics or diagrams are produced by
LibreOffice Draw. Source files are converted on the fly to pdf which are
then included in the source R markdown document
}

