% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocuskeletoncreator.R
\name{create_docu_package}
\alias{create_docu_package}
\title{Create a documentation package}
\usage{
create_docu_package(psPkgName, psPkgPath = ".", psDescTitle = NULL,
  psDescAuthor = NULL, psDescMaintainer = NULL, psDescDescription = NULL,
  psDescLicense = NULL)
}
\arguments{
\item{psPkgName}{name of the package}

\item{psPkgPath}{path below which package is created}

\item{psDescTitle}{title of the package}

\item{psDescAuthor}{author information}

\item{psDescDescription}{Description field in dcf}

\item{psDescLicense}{license information}
}
\description{
\code{create_docu_package} creates a documentation package using
\code{devtools::create}. Description information can be provided
by the additional arguments to the function. Changes to the
DESCRIPTION file are moved to a separate function called
\code{add_dcf_info}. If any of the arguments specific to the
DESCRIPTION file are not null, then the function
\code{add_dcf_info} is called.
}

