% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocuskeletoncreator.R
\name{create_docu_skeleton}
\alias{create_docu_skeleton}
\title{Create a new rmd document in source directory "rmd"}
\usage{
create_docu_skeleton(psDocuName, psPath = ".", pbEdit = FALSE)
}
\arguments{
\item{psDocuName}{name of the new document}

\item{psPath}{parent directory of rmd source dir}
}
\description{
In case it does not exist, the root document source directory
"rmd" is created below the parent directory psPath. All pdf
documents in the directory "rmd" are set to be ignored by git.
The draft document is created by \code{rmarkdown::draft}.
}

