% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocuodgdiagramcreator.R
\name{create_empty_odgdiagram}
\alias{create_empty_odgdiagram}
\title{Initialize an empty Odg diagram}
\usage{
create_empty_odgdiagram(psOdgDiagramName, psOdgDiagramPath = "vignettes/odg",
  psOdgTemplate = "odg_figure", psTemplatePkg = "rqudocuhelper",
  pbEdit = FALSE)
}
\arguments{
\item{psOdgDiagramName}{Name of odg diagram file}

\item{psOdgDiagramPath}{Path to diagram directory}

\item{psOdgTemplate}{Name of template to be used}

\item{psTemplatePkg}{Package from where template is used}

\item{pbEdit}{Should file be opened}
}
\description{
\code{create_empty_odgdiagram} uses function
\code{rmarkdown::draft} to initialize an empty
Odg diagram file.
}

