% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocudcfrefclass.R
\docType{class}
\name{DcfRefClass-class}
\alias{DcfRefClass}
\alias{DcfRefClass-class}
\title{Reference class for DESCRIPTION information}
\description{
\code{DcfRefClass} objects can be used to read
dcf files, to change information of the dcf
components and to write the dcf content back
to a file.
}
\section{Methods}{

\describe{
\item{\code{getAuthor()}}{getter for author}

\item{\code{getDescription()}}{getter for description}

\item{\code{getLicence()}}{getter for license}

\item{\code{getTitle()}}{getter for title}

\item{\code{readDcf(psDcfFile = "DESCRIPTION")}}{Read a description file using read.dcf}

\item{\code{setAuthor(psAuthor)}}{setter for author}

\item{\code{setDescription(psDescription)}}{setter for description}

\item{\code{setLicense(psLicense)}}{setter for license}

\item{\code{setTitle(psTitle)}}{setter for title}

\item{\code{writeDcf(psDcfFile = "DESCRIPTION")}}{Write a description file using write.dcf}
}}

