% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocugraphicsconverter.R
\name{insertOdgFigureAsPdf}
\alias{insertOdgFigureAsPdf}
\title{Inserts an odg draw figure into a rmarkdown text}
\usage{
insertOdgFigureAsPdf(psOdgFileStem, psOdgDir = "odg", psFigOutDir = ".")
}
\arguments{
\item{psOdgFileStem}{stem of odg figure file}

\item{psOdgDir}{directory where odg figure file is stored}

\item{psFigOutDir}{directory where output should be placed}
}
\description{
\code{insertOdgFigure} takes the name of a file containing a figure
in odg format, converts the content of that file into pdf using
function \code{convertLibOToPdf} and outputs the string in markdown
format to include the figure
}

