% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocudocustatusr6class.R
\docType{data}
\name{R6ClassDocuStatus}
\alias{R6ClassDocuStatus}
\title{R6 Class Representing Document Status Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6ClassDocuStatus
}
\description{
R6 Class Representing Document Status Objects
}
\keyword{datasets}

