% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocudocustatusr6class.R
\docType{data}
\name{R6ClassGenericTable}
\alias{R6ClassGenericTable}
\title{R6Class Representing A Generic Table Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6ClassGenericTable
}
\description{
R6Class Representing A Generic Table Object
}
\keyword{datasets}

