% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocudocustatusr6class.R
\docType{data}
\name{R6ClassVersion}
\alias{R6ClassVersion}
\title{R6 Class Representing Version Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6ClassVersion
}
\description{
Version strings can be separated into three parts
\enumerate{
  \item major version
  \item minor version
  \item fix level
}
For each of the three parts a private field is created
}
\keyword{datasets}

