% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocutablerefclass.R
\docType{class}
\name{refObjTable-class}
\alias{refObjTable}
\alias{refObjTable-class}
\title{Reference class to represent table-like reference objects}
\description{
A reference object of reference class \code{refObjTable} assumes
that a table simply consists of a tabel header and a table body.
Those components are represented by the fields of the reference class.
Apart from the initialisation method, the getter and setter methods
for the table header, we have a method to add additional rows to
the existing table body. The two methods \code{to_knitr_kable} and
\code{to_pander_pandoc} write tables in markdown format using
functions \code{knitr::kable()} and \code{pander::pandoc.table()}
}
\section{Fields}{

\describe{
\item{\code{sTableHeader}}{vector of table headers}

\item{\code{lTableBody}}{list of table body}
}}
\section{Methods}{

\describe{
\item{\code{addRow(plTableRow)}}{Adding a row represented by a list to the
 body of the table. The list representing
 the row must have the same names as the
 existing table body, otherwise, the row is
 not added to the table body.}

\item{\code{getTableHeader()}}{Getter for table header}

\item{\code{initialize()}}{Initialisation of table with empty body.}

\item{\code{setTableHeader(psTableHeader)}}{Setter for table header}

\item{\code{to_knitr_kable()}}{Output current table in markdown format using function
 knitr::kable(). In case the length of the specified table
 header is consistent with the number of columns, then
 the table header is added as columnnames of the data.frame
 representation of the table body.}

\item{\code{to_pander_pandoc(psStyle = "rmarkdown", psJustify = NULL)}}{Output current table in markdown format using the function
 pander::pandoc.table(). This method accepts two parameters
 psStyle and psJustify which are passed to
 to pander::pandoc.table().}
}}

