% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocuhelper.R
\name{render_scrreprt}
\alias{render_scrreprt}
\title{Render a document of documentclass scrreprt}
\usage{
render_scrreprt(psInput, psTemplateFile, plIncludes, pbViewResult = FALSE)
}
\arguments{
\item{psInput}{input Rmd file to be rendered}

\item{psTemplateFile}{template file given to pandoc}

\item{plIncludes}{list of named includes}

\item{pbViewResult}{flag whether to directly show result}
}
\description{
This function \code{render_scrreprt} is just
a wrapper for \code{rmarkdown::render}.
}

