% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocuskeletoncreator.R
\name{rmd_draft}
\alias{rmd_draft}
\title{Custom local copy of rmarkdown::draft}
\usage{
rmd_draft(file, template, package = NULL, create_dir = "default",
  pbOverwrite = FALSE)
}
\arguments{
\item{file}{name of the new document}

\item{template}{name of the template}

\item{package}{package where template can be found}

\item{create_dir}{whether or not to create a new directory for this document}

\item{pbOverwrite}{should existing files be overwritten}
}
\description{
\code{rmd_draft} corresponds to a local copy of
\code{rmarkdown::draft}. In contrast to the original
version, this version allows for the use of templates
with skeleton files which are already found in the
target directory.
}

