% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocusectioncountrefclass.R
\docType{class}
\name{SectionCount-class}
\alias{SectionCount-class}
\alias{sectionCount}
\title{reference Class for section counting}
\description{
reference Class for section counting
}

