% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocuodgdiagramcreator.R
\name{simple_draft}
\alias{simple_draft}
\title{Stripped down version of rmarkdown::draft}
\usage{
simple_draft(file, template, package = NULL, create_dir = FALSE)
}
\arguments{
\item{file}{name of resulting file}

\item{template}{name of template}

\item{package}{name of package where template is}

\item{create_dir}{should directory be created}
}
\description{
In contrast to \code{rmarkdown::draft}, we are here just
copying the files and subdirectories from the skeleton
subdirectory in the place that is given by file. We are
not doing anything with YAML here and we also want to
be independent of Rmd file endings.
}

