% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocucleanuptools.R
\name{cleanup_vignettes}
\alias{cleanup_vignettes}
\title{Cleaning up a vignette directory}
\usage{
cleanup_vignettes(psPath = "vignettes", psPattern = "pdf$")
}
\arguments{
\item{psPath}{path to root directory which should be cleaned}

\item{psPattern}{that files to be clean match}
}
\description{
\code{cleanup_vignettes} removes all output files that were
produced by compiling document sources. The goal is to have
only source files after the cleanup process.
}
\details{
The approach taken in function \code{cleanup_vignettes} is
to specify the pattern of the files to be deleted and the
pattern of the files to keep. The reason for this is that
the number of ouput formats to be specified is much smaller
than the number of files that must be kept. The files to be
removed are first collected into a vector, then the user
is asked whether to remove all files. This is a security
feature to avoid deleting files accidentally.
}

