% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocusectioncountrefclass.R
\docType{class}
\name{SectionEnumerator-class}
\alias{SectionEnumerator}
\alias{SectionEnumerator-class}
\title{Reference class for automatically enumerating section titles}
\description{
Reference class for automatically enumerating section titles
}

