% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqudocusectioncountrefclass.R
\docType{class}
\name{SectionEnumerator-class}
\alias{SectionEnumerator}
\alias{SectionEnumerator-class}
\title{Reference class for automatically enumerating section titles}
\description{
The level of a Markdown section headers is indicated using
hashes in front of the title. YAML frontmatter allows for
to specify section header numbering, but it is not clear
how to mix headers with and without numbers in the same
document. The \code{SectionEnumerator} reference class allows
us to specify section header numbers only for certain titles.
}
\details{
The main method of a SectionEnumerator reference object
is \code{displayNumSection}. This takes as input a string
of an unnumbered section, parses that string and computes
the section number based on the number of hashes.
}
\section{Methods}{

\describe{
\item{\code{displayNumSection(psUnNumSection = NULL)}}{Parsing of unnumbered section string and display the string including the section number}

\item{\code{initialize()}}{Initalisation of object fields}

\item{\code{parseUnNumSection()}}{Parse the unnumbered section string and assign the object fields}

\item{\code{setUnNumSection(psUnNumSection)}}{Setter for unnumbered section string}
}}

